# backGO/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple user service",
        "title": "User API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/minigame/finish": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MiniGame"
                ],
                "summary": "Завершить мини-игру",
                "parameters": [
                    {
                        "description": "Game Result",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.FinishMiniGameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/minigame/start": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MiniGame"
                ],
                "summary": "Начать мини-игру",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.StartMiniGameResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.FinishMiniGameRequest": {
            "type": "object",
            "properties": {
                "hash": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "session_id": {
                    "type": "string"
                }
            }
        },
...
        "handlers.StartMiniGameResponse": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "session_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}